requirements.txt
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
black==25.1.0
certifi==2025.1.31
click==8.1.8
Django==5.1.5
django-ckeditor==6.7.2
django-js-asset==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
flake8==7.1.1
googletrans==4.0.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.4
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
